Move verify_valid_installtree to source module utils (#1844287)
It's used only by modules now. We can safely move it.

Related: rhbz#1844287
Related: rhbz#1849093
