+++ pyanaconda/payload/image.py
@@ -28,7 +28,6 @@ import blivet.arch
 from blivet.size import Size
 
 from pyanaconda import isys
-from pyanaconda.core.util import join_paths
 from pyanaconda.errors import errorHandler, ERROR_RAISE, InvalidImageSizeError, MissingImageError
 from pyanaconda.modules.common.constants.objects import DEVICE_TREE
 from pyanaconda.modules.common.constants.services import STORAGE
@@ -127,21 +126,6 @@ def find_first_iso_image(path, mount_path="/mnt/install/cdimage"):
         return fn
 
     return None
-
-
-def verify_valid_installtree(path):
-    """Check if the given path is a valid installtree repository.
-
-    :param str path: install tree path
-    :returns: True if repository is valid false otherwise
-    :rtype: bool
-    """
-    repomd_path = join_paths(path, "repodata/repomd.xml")
-
-    if os.path.exists(repomd_path) and os.path.isfile(repomd_path):
-        return True
-
-    return False
 
 
 def _check_repodata(mount_path):
